work around cors issue using local proxy to load project files
to load scratch projects in a fork of the scratch-vm or scratch-gui, we need to get the project file from api.scratch.mit.edu but that server doesn't accept requests from localhost or any origin other than scratch.mit.edu. To works around this we need a server side script like this that acts as a CORS proxy.

https://github.com/LLK/scratch-gui/issues/8723#event-8171009287